C++ and Rust int32->int64

+++ prime-number/c++/cmd.cpp
@@ -3,17 +3,17 @@
 
 int main(int argc, char *argv[])
 {
-    int32_t primeNumberCount = 100;
+    int64_t primeNumberCount = 100;
     if (argc == 2) {
         primeNumberCount = atoi(argv[1]);
     }
 
-    int32_t number = 0;
+    int64_t number = 0;
     while (primeNumberCount > 0) {
         number++;
 
-        int32_t j = 0;
-        for (int32_t i = 1; i <= number; i++) {
+        int64_t j = 0;
+        for (int64_t i = 1; i <= number; i++) {
             if (number % i == 0) {
                 j++;
             }
